+++ resources/migrations/20191023162400-unavailable-roles-for-api-key.up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE api_key ADD COLUMN unavailableRoles jsonb;