+++ resources/sql/queries.sql
@@ -472,12 +472,16 @@ INSERT INTO application_event (appId, eventData)
 VALUES (:application, :eventdata::jsonb);
 
 -- :name add-api-key! :insert
-INSERT INTO api_key (apiKey, comment)
-VALUES (:apikey, :comment)
+INSERT INTO api_key (apiKey, comment, unavailableRoles)
+VALUES (
+:apikey,
+:comment,
+:unavailableroles::jsonb
+)
 ON CONFLICT DO NOTHING;
 
 -- :name get-api-key :? :1
-SELECT apiKey FROM api_key
+SELECT apiKey, unavailableRoles::TEXT FROM api_key
 WHERE apiKey = :apikey;
 
 -- :name get-application-by-invitation-token :? :1