+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Changes since v2.6
   - `/api/email/send-reminders` sends all of the above emails.
 - Allow users to change their email address, in case the identity provider
   doesn't provide an email address or the users want to use a different one (#1884) 
+- Allow API keys to have unavailable roles (#1662)
 
 ### Enhancements
 - Application search tips hidden behind question mark icon (#1767)